Unify protocols between client and server

+++ server/common/rule.go
@@ -6,7 +6,7 @@ package common
 //
 // Suppose you need 100 exp until level 2 and 200 exp until level 3. If you
 // have 160 exp, your level is 2, levExp is 100, curExp is 60, ratio is 0.6.
-func Exp2Level(exp int64) (level int, levExp, curExp int64, ratio float64) {
+func Exp2Level(exp int64) (level int32, levExp, curExp int64, ratio float64) {
 	for i, needUntil := range expTable {
 		if i == 0 {
 			continue
@@ -14,7 +14,7 @@ func Exp2Level(exp int64) (level int, levExp, curExp int64, ratio float64) {
 		needBefore := expTable[i-1]
 
 		if exp < needUntil {
-			level = i
+			level = int32(i)
 			levExp = needUntil - needBefore
 			curExp = exp - needBefore
 			ratio = float64(curExp) / float64(levExp)
@@ -23,7 +23,7 @@ func Exp2Level(exp int64) (level int, levExp, curExp int64, ratio float64) {
 	}
 
 	// has exp above top level
-	level = len(expTable)
+	level = int32(len(expTable))
 	levExp = 0
 	curExp = 0
 	ratio = 0
@@ -32,7 +32,7 @@ func Exp2Level(exp int64) (level int, levExp, curExp int64, ratio float64) {
 
 // Level2Sight returns a radius of user's sight derived from user's level.
 // Returned radius is METER.
-func Level2Sight(level int) float64 {
+func Level2Sight(level int32) float64 {
 	return 10.0 + float64(level)*10.0
 }
