+++ server/common/rule_test.go
@@ -7,7 +7,7 @@ import (
 func TestExpBoundary(t *testing.T) {
 	type expTestSet struct {
 		exp      int64
-		level    int
+		level    int32
 		residual int64
 	}
 
@@ -17,9 +17,7 @@ func TestExpBoundary(t *testing.T) {
 			continue
 		}
 		samples = append(samples, expTestSet{
-			expTable[i-1],
-			i,
-			0,
+			expTable[i-1], int32(i), 0,
 		})
 	}
 
@@ -35,7 +33,7 @@ func TestExpBoundary(t *testing.T) {
 func TestLargeExp(t *testing.T) {
 	maxInt := int64(^uint64(0) >> 1)
 	level, levExp, curExp, ratio := Exp2Level(maxInt)
-	if level != len(expTable) || levExp != 0 || curExp != 0 || ratio != 0 {
+	if level != int32(len(expTable)) || levExp != 0 || curExp != 0 || ratio != 0 {
 		t.Errorf("Exp2Level(%v) = %v, %v, %v, %v; want %v, %v, %v, %v",
 			maxInt, level, levExp, curExp, ratio, len(expTable), 0, 0, 0)
 	}
